:recycle: プロバイダーディレクトリを作成し、42認証のテーブルを作成する関数を移動
各プロバイダー対応できるようにするため
