+++ backend/pong/oauth2/providers/__init__.py
@@ -1,0 +1,1 @@
+from . import forty_two_authorization
